Extract getData function
