+++ bin/index.js
@@ -1,73 +1,18 @@
 #!/usr/bin/env node
 "use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
 var program = require('commander').program;
-var axios = require('axios');
-var url = require('url');
-var _a = require('./command/latest'), LATEST_URL = _a.LATEST_URL, processCommands = _a.processCommands;
-var BASE_URL = new URL(LATEST_URL);
-var getData = function (url) { return __awaiter(void 0, void 0, void 0, function () {
-    var response, data, error_1;
-    return __generator(this, function (_a) {
-        switch (_a.label) {
-            case 0:
-                _a.trys.push([0, 2, , 3]);
-                return [4, axios.get(url)];
-            case 1:
-                response = _a.sent();
-                data = response.data;
-                console.table(data.rates);
-                return [3, 3];
-            case 2:
-                error_1 = _a.sent();
-                console.log(error_1);
-                return [3, 3];
-            case 3: return [2];
-        }
-    });
-}); };
+var processCommands = require('./command/latest').processCommands;
 program
-    .option('-b, --base <type>', 'change the base currency')
-    .option('-s, --symbols <type>', 'list preferred output currencies')
-    .option('-a, --amount <type>', 'the amount to be converted')
-    .option('-p, --places <type>', 'round numbers to decimal place')
-    .option('--source <type>', 'switch source data between forex or bank view');
+    .option('-b, --base <symbol>', 'change the base currency')
+    .option('-s, --symbols [symbols]', 'list preferred output currencies')
+    .option('-a, --amount <amount>', 'the amount to be converted')
+    .option('-p, --places <decimals>', 'round numbers to decimal place')
+    .option('--source <source>', 'switch source data between forex or bank view')
+    .action(function (option) {
+    processCommands(option);
+});
+program
+    .command('convert')
+    .alias('c')
+    .action(function () { return console.log('convert'); });
 program.parse(process.argv);
-var PARSED_BASE_URL = processCommands(BASE_URL, program);
-console.log(PARSED_BASE_URL.href);
-getData(PARSED_BASE_URL.href);
